Pridany deletePost pro adminy, opraven bug s ID v loginu, a settings, opraven profile kde se nenacitali nektery informace, pridano admin odkaz do headeru na admin panel

+++ client/profiles.php
@@ -132,13 +132,14 @@ foreach ($dataPosts as $post) {
     echo '        </div>';
    
 
-    if ($_SESSION['admin'] == 1) {
-        $currentUrl = "http://$_SERVER[HTTP_HOST]$_SERVER[REQUEST_URI]";
-        $_SESSION['redirectBackUrl'] = $currentUrl;
-        echo $currentUrl; 
-        echo '<button class="bg-red-500 hover:bg-red-700 text-white font-bold">';
-        echo '<a href="../admin/deletepost.php?id=' . $post['postID'] . '">Delete post</a>';
-        echo '</button>';
+    if(isset($_SESSION['admin'])){
+        if($_SESSION['admin'] == 1){
+            $currentUrl = "http://$_SERVER[HTTP_HOST]$_SERVER[REQUEST_URI]";
+            $_SESSION['redirectBackUrl'] = $currentUrl;
+            echo '<button class="bg-red-500 hover:bg-red-700 text-white font-bold">';
+            echo '<a href="../admin/deletepost.php?id=' . $post['postID'] . '">Delete post</a>';
+            echo '</button>';
+        }
     }
 }
